+++ oscil~.c
@@ -57,7 +57,7 @@ typedef struct _oscil
 
 } t_oscil;
 
-void *oscil_new(t_symbol *s, short argc, t_atom *argv);
+void *oscil_new(t_symbol *s, int argc, t_atom *argv);
 static t_int *oscil_perform(t_int *w);
 void oscil_dsp(t_oscil *x, t_signal **sp);
 void build_waveform(t_oscil *x);
@@ -68,13 +68,13 @@ void oscil_sawtooth(t_oscil *x);
 void oscil_square(t_oscil *x) ;
 void oscil_triangle(t_oscil *x);
 void oscil_buzz(t_oscil *x );
-void oscil_list (t_oscil *x, t_symbol *msg, short argc, t_atom *argv);
+void oscil_list (t_oscil *x, t_symbol *msg, int argc, t_atom *argv);
 void oscil_fadetime (t_oscil *x, t_floatarg fade_ms) ;
 void oscil_fadetype(t_oscil *x, t_floatarg ftype);
 void oscil_harmcount(t_oscil *x, t_floatarg harms);
 void oscil_interpolate(t_oscil *x, t_floatarg tog);
 void oscil_dsp_free(t_oscil *x);
-void oscil_amph(t_oscil *x, t_symbol *msg, short argc, t_atom *argv);
+void oscil_amph(t_oscil *x, t_symbol *msg, int argc, t_atom *argv);
 t_int *oscil_perform_interpolate(t_int *w);
 
 
@@ -100,7 +100,7 @@ void oscil_tilde_setup(void)
 }
 
 
-void oscil_list (t_oscil *x, t_symbol *msg, short argc, t_atom *argv)
+void oscil_list (t_oscil *x, t_symbol *msg, int argc, t_atom *argv)
 {
   short i;
   int harmonic_count = 0;
@@ -113,7 +113,7 @@ void oscil_list (t_oscil *x, t_symbol *msg, short argc, t_atom *argv)
   build_waveform(x);
 }
 
-void oscil_amph(t_oscil *x, t_symbol *msg, short argc, t_atom *argv)
+void oscil_amph(t_oscil *x, t_symbol *msg, int argc, t_atom *argv)
 {
   short i;
   int harmonic_count = 0;
@@ -183,7 +183,7 @@ void oscil_interpolate(t_oscil *x, t_floatarg flag)
 }
 
 
-void *oscil_new(t_symbol *s, short argc, t_atom *argv)
+void *oscil_new(t_symbol *s, int argc, t_atom *argv)
 {
   float init_freq;
   t_symbol *init_waveform_symbol;
